# Conflicts:
#	gradle.properties
#	settings.gradle
#	src/main/java/com/gregtechceu/gtceu/api/capability/recipe/EURecipeCapability.java
#	src/main/java/com/gregtechceu/gtceu/api/registry/GTRegistries.java
#	src/main/java/com/gregtechceu/gtceu/client/forge/ForgeClientEventListener.java
#	src/main/java/com/gregtechceu/gtceu/common/capability/EnvironmentalHazardSavedData.java
#	src/main/java/com/gregtechceu/gtceu/common/network/GTNetwork.java
#	src/main/java/com/gregtechceu/gtceu/forge/ForgeCommonEventListener.java
#	src/main/resources/gtceu.mixins.json

+++ src/main/java/com/gregtechceu/gtceu/api/material/material/Material.java
@@ -1082,7 +1082,7 @@ public class Material implements Comparable<Material> {
             return this;
         }
 
-        public Builder radioactiveHazard(int multiplier) {
+        public Builder radioactiveHazard(float multiplier) {
             properties.setProperty(HAZARD, new HazardProperty(HazardProperty.HazardTrigger.ANY,
                     GTMedicalConditions.CARCINOGEN, multiplier, true));
             return this;
